adding work on ECM surface
Results are in generated by generateECMSurface.m and are contained in finalOutput.txt. Results for a single pulse are calculated in exampleFitting. Some data cleaning may be required since there are huge values for R1 and C1 sometimes or they are negative. Constrained optimization doesn't really fix the issue—I believe the pulse data is just bad. For example see run24 in -20C data.